+++ src/cli/mod.rs
@@ -2,4 +2,5 @@ pub mod connect;
 pub mod set_key;
 pub mod stats;
 pub mod stop;
+pub mod up;
 pub mod update;